+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### VB -> C#
 
+* Convert with blocks using structs to a local ref variable[#634](https://github.com/icsharpcode/CodeConverter/issues/634)
 
 ### C# -> VB